split all publishing-related functionality out of modals.js and into a new file, js/fc/ui/publish.js.
